# EDIT - SignerPool 개선
- Producer - Consumer 모델로 변경해서 lock 때문에 오는 병목현상을 줄임

+++ src/plugins/block_producer_plugin/block_producer_plugin.cpp
@@ -71,7 +71,7 @@ private:
     bitset<256> id_bits = getOptimalMergerId(blocks, producersCount);
 
     bitset<256> my_id(app().getId());
-    int dist = getHammingDistance(id_bits, my_id);
+    float dist = getHammingDistance(id_bits, my_id);
 
     return dist;
   }
